Keep validation on message queue sizes
We need this in case the incentivized channel is full: if we don't have
this check, we'll burn the fee and fail to send the message.

For the outbound channel we're fine without this check for now, but I'd
rather keep it as a preventative measure. If we add any mutations to the
try_mutate closure in the future and the queue ends up being full, we'll
trigger the mutation without appending the message to the queue. This is
assuming the mutations are added before appending the message to the
queue.