Permit non-consecutive increases in validator set
Changes the validator set successions to still be sequential, but non-consecutive.

For bridge security against collusion by validators, it's important that
these validators are still bonded, but the timing of the signature itself
is secondary.

As such, even if some validators have rotated out, this change permits
keeping the bridge alive so long as 2/3rds of the `currentValidatorSet`
are still bonded and sign the commitment.

Co-authored-by: bhargavbh <2bhargav5@gmail.com>
Co-authored-by: Alistair Stewart <alistair@web3.foundation>