ensure same block used for ancestry proof & header
For the proof to be accepted, it's crucial that the payload (MMR root)
contained in the header matches the ancestry proof's best MMR root. If
the best block happened to get updated between both calls, this may not
be the case. Ergo: use the block hash of the header provided in the
equivocation proof for the ancestry proof generation call.

Co-authored-by: Ron <yrong1997@gmail.com>